Swap out segformers
