+++ surya/settings.py
@@ -35,14 +35,13 @@ class Settings(BaseSettings):
     @computed_field
     def TORCH_DEVICE_DETECTION(self) -> str:
         if self.TORCH_DEVICE is not None:
-            # Does not work with mps
-            if "mps" in self.TORCH_DEVICE:
-                return "cpu"
-
             return self.TORCH_DEVICE
 
         if torch.cuda.is_available():
             return "cuda"
+
+        if torch.backends.mps.is_available():
+            return "mps"
 
         # Does not work with mps
         return "cpu"
